+++ jest.config.js
@@ -2,6 +2,7 @@ module.exports = {
   testEnvironment: "jsdom",
   setupFiles: [
     "./Vimkit Extension/js/mocks.js",
+    "./Vimkit Extension/js/lib/svim-scripts.js",
     "./Vimkit Extension/js/settings.js",
     "./Vimkit Extension/js/WebExtensionCommunicator.js",
     "./Vimkit Extension/js/command-dispatcher.js",